requirements.txt
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
joblib>=1.3.0
pyyaml>=6.0
huggingface-hub>=0.19.0
matplotlib>=3.7.0
seaborn>=0.12.0
